Moved DB logic out of resolvers, built out some wrapper classes
Thos should open up possibilities of other DB types in the future, as long as the function names are the same (JS doesn't seem to have a good way to handle Interfaces it seems...)

This also simplifieds a lot of the resolvers logic. SmashGG API calls might want to be abstracted out more too for the same reason.